Fixed the deploy shell bug

+++ bin/deploy.sh
@@ -1,17 +1,24 @@
 #!/bin/bash
 
-su - jesus
+user = `whoami`
 
-cd /home/jesus/developkit/repository/git/jtodo
+if [ $user = "jesus" ]; then
 
-# update the code from git
-git pull -r
+    cd /home/jesus/developkit/repository/git/jtodo
+    
+    # update the code from git
+    git pull -r
 
-cd /home/jesus/developkit/service/tomcat/jtodo/t1/bin
-./shutdown.sh
-
-# copy the execution jar file to the tomcat
-rm -rf /home/jesus/developkit/service/tomcat/jtodo/t1/webapps/*
-cp /home/jesus/developkit/repository/git/jtodo/src/web/target/jtodo*.war /home/jesus/developkit/service/tomcat/jtodo/t1/webapps/jtodo.war
-
-./start.sh
+    if [ $? -ne 0]; then
+        cd "/home/jesus/developkit/service/tomcat/jtodo/t1/bin"
+        ../shutdown.sh
+        
+        rm -rf /home/jesus/developkit/service/tomcat/jtodo/t1/webapps/*
+        cp /home/jesus/developkit/repository/git/jtodo/src/web/target/jtodo*.war /home/jesus/developkit/service/tomcat/jtodo/t1/webapps/jtodo.war
+            
+        cd /home/jesus/developkit/service/tomcat/jtodo/t1/bin
+        ../start.sh
+    else
+        echo "Git pull -r failed"
+else
+    echo "Please execute as jesus"
